added task only deleted by their own user + only admin can access user managing pages

+++ src/AppBundle/Controller/TaskController.php
@@ -83,6 +83,7 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted('delete', $task);
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
